Renamed SafeClassTagMacro -> ConcreteTagMacro; deriving unions/intersections from givens; added tests
